feat: create comment backend
