+++ backend/midias/comment/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Comment
+
+# Register your models here.
+admin.site.register(Comment)
