+++ backend/midias/comment/models.py
@@ -1,0 +1,17 @@
+from django.db import models
+
+from review.models import Review
+from user.models import User
+
+
+
+
+# Create your models here.
+class Comment(models.Model):
+    content = models.TextField()
+    review = models.ForeignKey(Review, on_delete=models.CASCADE, related_name="comments")
+    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="comments")
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return self.content
