+++ backend/midias/comment/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from .models import Comment
+
+
+class CommentSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Comment
+        fields = "__all__"
