+++ backend/midias/review/models.py
@@ -25,13 +25,3 @@ class Review(models.Model):
 
     def __str__(self):
         return self.title
-
-
-class Comment(models.Model):
-    content = models.TextField()
-    review = models.ForeignKey(Review, on_delete=models.CASCADE, related_name="comments")
-    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="comments")
-    created_at = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return self.content
